fix: ui/ux login signup and profile
