+++ src/hooks/useCurrentUser.jsx
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { UserContext } from '../contexts/UserContext'
+
+export default () => useContext(UserContext)
