+++ src/navigation/NavigationTab.jsx
@@ -3,13 +3,13 @@ import { StyleSheet } from 'react-native'
 import { Fontisto } from '@expo/vector-icons'
 import { COLORS, SPACING } from '../utils/theme'
 import { HomeScreen } from '../screens/HomeScreen'
-import { ProfileScreen } from '../screens/profile/ProfileScreen'
+import { ProfileScreen } from '../screens/ProfileScreen'
 import { FavoriteScreen } from '../screens/FavoriteScreen'
 const Tab = createBottomTabNavigator()
 
 const TAB_ICON = {
   Home: 'home',
-  Profile: 'person'
+  Profile: 'person',
   Favorites: 'bookmark-alt'
 }
 const screenOptions = ({ route }) => {
@@ -30,8 +30,8 @@ export function NavigationTab () {
   return (
     <Tab.Navigator screenOptions={screenOptions}>
       <Tab.Screen name='Home' component={HomeScreen} />
-      <Tab.Screen name='Profile' component={ProfileScreen} />
       <Tab.Screen name='Favorites' component={FavoriteScreen} />
+      <Tab.Screen name='Profile' component={ProfileScreen} />
     </Tab.Navigator>
   )
 }
