+++ src/screens/AuthScreen/AuthScreen.styles.js
@@ -4,7 +4,6 @@ export const styles = StyleSheet.create({
   container: {
     justifyContent: 'center',
     flex: 1,
-    padding: 16,
     backgroundColor: '#fff',
     marginTop: StatusBar.currentHeight
   }
