+++ src/screens/UserScreen/index.jsx
@@ -4,25 +4,27 @@ import { styles } from './UserInfoScreen.styles'
 import { UserContext } from '../../contexts/UserContext'
 export const UserInfoScreen = () => {
   const { currentUser, setCurrentUser } = useContext(UserContext)
-  console.log(currentUser)
+
   const handleLogout = () => {
     setCurrentUser(null)
   }
+
   return (
     <ScrollView style={styles.container}>
       <View style={styles.header}>
         <Image
           style={styles.profileImage}
-          source={require('../../../assets/images/avatar.png')}
+          source={{ uri: `https://robohash.org/${currentUser.username}` }}
         />
         <View style={styles.profileInfo}>
-          <Text style={styles.profileEmail}>{currentUser}</Text>
+          <Text style={styles.profileUsername}>{currentUser.username}</Text>
+
+          <Text style={styles.profileEmail}>{currentUser.email}</Text>
         </View>
         <TouchableOpacity style={styles.button} onPress={handleLogout}>
           <Text style={styles.buttonText}>Salir</Text>
         </TouchableOpacity>
       </View>
-
     </ScrollView>
   )
 }
