--- src/screens/profile/ProfileScreen.jsx
@@ -1,22 +1,0 @@
-import React, { useContext } from 'react'
-import { UserContext } from '../../contexts/UserContext'
-import { UserInfoScreen } from '../user/UserInfoScreen'
-// import { SigninScreen } from '../signin/SigninScreen'
-import { AuthScreen } from '../auths/AuthScreen'
-
-export const ProfileScreen = () => {
-  const { currentUser } = useContext(UserContext)
-
-  return (
-    <>
-      {currentUser
-        ? (
-          <UserInfoScreen />
-          )
-        : (
-          <AuthScreen />
-          )}
-    </>
-
-  )
-}
